python/gitlab-reg-tag-cleaner/requirements.txt
pyyaml
python-gitlab